Assignment select + edit now available as modals
commit c3ef2149f3175ad4b7f707ce56e62be3cfd63f35
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Tue Feb 11 19:40:42 2025 +0200

    modal assignment edit and select now working

commit 1136b7b8437d248b5994ff5fa084dead2cadb374
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Tue Feb 11 14:31:52 2025 +0200

    fixing editor-in-editor bug

commit 3cf61e751ae11008e4b7f356eb024b96654d9b69
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Tue Feb 11 11:34:26 2025 +0200

    formatting

commit 037a61a9b60d6eb08cbc2aa1f5ca738f47a87200
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Tue Feb 11 11:22:43 2025 +0200

    debug buttons abstracted

commit f8e2a374e16cad5740c723fcb316c36e434ce558
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Mon Feb 10 18:37:22 2025 +0200

    cleaning up assignment edit for modal solution

commit d78cd807c441cd3a13a6e009cf045ad0888e7481
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Mon Feb 10 16:10:48 2025 +0200

    assignment edit also available as modal

commit 3e4a4a6dcab923cea5ba0898822c0a13d89b77c9
Author: eliassyyrilaLUT <elias.syyrila@lut.fi>
Date:   Mon Feb 10 13:25:57 2025 +0200

    assignment selector now available as a modal component
