+++ src/components/Context.tsx
@@ -1,5 +1,4 @@
 import { createContext, useState } from "react";
-import { functionResultToSnackBar, SnackBarAttributes } from "./SnackBarComp";
 import {
   CodeAssignmentData,
   CodeAssignmentDatabase,
@@ -7,6 +6,7 @@ import {
   ModuleData,
   SetData,
 } from "../types";
+import { functionResultToSnackBar, SnackBarAttributes } from "./SnackBarComp";
 
 export const UIContext = createContext(null);
 export const ActiveObjectContext = createContext(null);
@@ -75,47 +75,34 @@ export const ActiveObjectProvider = ({ children }: { children: any }) => {
     useState<Array<CodeAssignmentDatabase>>(undefined);
   const [activeModule, setActiveModule] = useState<ModuleData>(null);
   const [activeSet, setActiveSet] = useState<SetData>(null);
-  const [selectAssignment, setSelectAssignment] = useState<boolean>(null);
   const [genericModuleAssignmentCount, setGenericModuleAssignmentCount] =
     useState<number>(null);
   const [tempAssignment, setTempAssignment] =
     useState<CodeAssignmentData>(undefined);
-  const [previousPath, setPreviousPath] = useState<string>(undefined);
 
   function handleActiveCourse(value: CourseData) {
     setActiveCourse(value);
   }
-
   function handleActivePath(value: string) {
     setActivePath(value);
   }
-
   function handleActiveAssignment(value: CodeAssignmentData) {
     setActiveAssignment(value);
   }
-
   function handleActiveModule(value: ModuleData) {
     setActiveModule(value);
   }
-
   function handleActiveSet(value: SetData) {
     setActiveSet(value);
   }
-
   function handleActiveAssignments(value: CodeAssignmentDatabase[]) {
     setActiveAssignments(value);
-  }
-  function handleSelectAssignment(value: boolean) {
-    setSelectAssignment(value);
   }
   function handleGenericModuleAssignmentCount(value: number) {
     setGenericModuleAssignmentCount(value);
   }
   function handleTempAssignment(value: CodeAssignmentData) {
     setTempAssignment(value);
-  }
-  function handlePreviousPath(value: string) {
-    setPreviousPath(value);
   }
 
   return (
@@ -134,14 +121,10 @@ export const ActiveObjectProvider = ({ children }: { children: any }) => {
           handleActiveModule,
           activeSet,
           handleActiveSet,
-          selectAssignment,
-          handleSelectAssignment,
           genericModuleAssignmentCount,
           handleGenericModuleAssignmentCount,
           tempAssignment,
           handleTempAssignment,
-          previousPath,
-          handlePreviousPath,
         }}
       >
         {children}
